# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Canada Soccer API",
    "version": "2.0.0",
    "description": "Open Canadian Premier League match, live, table, player, team, advanced-stat, referee, venue, weather, and odds data from 2019 through the current season.",
    "license": { "name": "CC BY 4.0", "url": "https://creativecommons.org/licenses/by/4.0/" }
  },
  "servers": [{ "url": "https://canadasoccerapi.com/api/v1", "description": "Production" }],
...
  "paths": {
...
    "/matches": {
      "get": {
        "tags": ["Matches"], "summary": "Query the complete match archive and live schedule",
        "parameters": [
          { "$ref": "#/components/parameters/Season" }, { "$ref": "#/components/parameters/Team" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["FINISHED", "UPCOMING", "LIVE", "IN_PLAY", "HALF_TIME", "POSTPONED", "CANCELLED", "SUSPENDED"] } },
          { "name": "match_id", "in": "query", "schema": { "type": "string" } },
          { "name": "date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "$ref": "#/components/parameters/From" }, { "$ref": "#/components/parameters/To" },
          { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Offset" }, { "$ref": "#/components/parameters/Order" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/match-stats": {
      "get": {
        "tags": ["Matches"], "summary": "Get match-level advanced team statistics and xG",
        "parameters": [
          { "name": "match_id", "in": "query", "description": "Required unless season is supplied", "schema": { "type": "string" } },
          { "$ref": "#/components/parameters/Season" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/odds-quotes": {
      "get": {
        "tags": ["Odds"], "summary": "Query timestamped bookmaker snapshots",
        "parameters": [
          { "$ref": "#/components/parameters/Season" }, { "$ref": "#/components/parameters/Team" },
          { "name": "bookmaker", "in": "query", "schema": { "type": "string" } },
          { "name": "market", "in": "query", "schema": { "type": "string" } },
          { "name": "date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Offset" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    }
  },
  "components": {
    "parameters": {
      "Season": { "name": "season", "in": "query", "schema": { "type": "integer", "minimum": 2019, "example": 2026 } },
      "Team": { "name": "team", "in": "query", "schema": { "type": "string", "example": "Forge" } },
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 100 } },
      "Offset": { "name": "offset", "in": "query", "schema": { "type": "integer", "minimum": 0, "default": 0 } },
      "Order": { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
      "From": { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
      "To": { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } }
    },
    "responses": {
      "Success": { "description": "Successful JSON response", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } },
      "Error": { "description": "Standard error response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Error": {
        "type": "object", "required": ["error", "message", "status"],
        "properties": { "error": { "type": "string" }, "message": { "type": "string" }, "status": { "type": "integer" } }
      }
    }
  }
}